feat: enhance error handling and response structure for IP and domain searches
